[CLEANUP] Move CSS parsing to a separate class
Part of #734 and #544.  Relates to #994.
