+++ tests/Unit/CssInlinerTest.php
@@ -7,6 +7,7 @@ namespace Pelago\Emogrifer\Tests\Unit;
 use Pelago\Emogrifier\CssInliner;
 use Pelago\Emogrifier\HtmlProcessor\AbstractHtmlProcessor;
 use Pelago\Emogrifier\Tests\Support\Traits\AssertCss;
+use Pelago\Emogrifier\Utilities\ParsedCss;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\CssSelector\CssSelectorConverter;
 use Symfony\Component\CssSelector\Exception\SyntaxErrorException;
@@ -3713,8 +3714,9 @@ final class CssInlinerTest extends TestCase
         $copyUninlinableCssToStyleNode->setAccessible(true);
 
         $domDocument = $subject->getDomDocument();
+        $parsedCss = new ParsedCss('');
 
-        $copyUninlinableCssToStyleNode->invoke($subject, '');
+        $copyUninlinableCssToStyleNode->invoke($subject, $parsedCss);
         $expectedHtml = $subject->render();
 
         $styleElement = $domDocument->getElementsByTagName('style')->item(0);
@@ -3722,7 +3724,7 @@ final class CssInlinerTest extends TestCase
         self::assertInstanceOf(\DOMElement::class, $styleElement->parentNode);
         $styleElement->parentNode->removeChild($styleElement);
 
-        $copyUninlinableCssToStyleNode->invoke($subject, '');
+        $copyUninlinableCssToStyleNode->invoke($subject, $parsedCss);
 
         self::assertSame($expectedHtml, $subject->render());
     }
